backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
structlog==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.44.0
vine==5.1.0
wcwidth==0.6.0
